tests/wpt/web-platform-tests/fullscreen/api/element-ready-check-iframe-child-manual.html

<!DOCTYPE html>
<title>Element ready check for child of iframe</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<iframe><!-- script inserts child here --></iframe>
<script>
async_test(t => {
  const div = document.createElement("div");
  document.querySelector("iframe").appendChild(div);

  document.onfullscreenchange = t.step_func_done(() => {
    assert_equals(document.fullscreenElement, div);
  });
  document.onfullscreenerror = t.unreached_func("fullscreenerror event");

  trusted_request(t, div, document.body);
});
</script>
